migrate to use extension panel adapter
