+++ packages/studio-base/src/panels/CallService/CallService.tsx
@@ -1,0 +1,247 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at http://mozilla.org/MPL/2.0/
+
+import { Button, TextField, Tooltip, Typography, inputBaseClasses } from "@mui/material";
+import { useCallback, useEffect, useMemo, useState } from "react";
+import { makeStyles } from "tss-react/mui";
+
+import Log from "@foxglove/log";
+import { PanelExtensionContext, SettingsTreeAction } from "@foxglove/studio";
+import Stack from "@foxglove/studio-base/components/Stack";
+import { Config } from "@foxglove/studio-base/panels/CallService/types";
+import { fonts } from "@foxglove/studio-base/util/sharedStyleConstants";
+
+import { defaultConfig, settingsActionReducer, useSettingsTree } from "./settings";
+
+const log = Log.getLogger(__dirname);
+
+type Props = {
+  context: PanelExtensionContext;
+};
+
+type State = {
+  status: "requesting" | "error" | "success";
+  value: string;
+};
+
+const useStyles = makeStyles<{ buttonColor?: string }>()((theme, { buttonColor }) => {
+  const augmentedButtonColor = buttonColor
+    ? theme.palette.augmentColor({
+        color: { main: buttonColor },
+      })
+    : undefined;
+
+  return {
+    button: {
+      backgroundColor: augmentedButtonColor?.main,
+      color: augmentedButtonColor?.contrastText,
+
+      "&:hover": {
+        backgroundColor: augmentedButtonColor?.dark,
+      },
+    },
+    textarea: {
+      height: "100%",
+
+      [`.${inputBaseClasses.root}`]: {
+        width: "100%",
+        height: "100%",
+        textAlign: "left",
+        backgroundColor: theme.palette.background.paper,
+        overflow: "hidden",
+        padding: theme.spacing(1, 0.5),
+
+        [`.${inputBaseClasses.input}`]: {
+          height: "100% !important",
+          lineHeight: 1.4,
+          fontFamily: fonts.MONOSPACE,
+          overflow: "auto !important",
+          resize: "none",
+        },
+      },
+    },
+  };
+});
+
+function parseInput(value: string): { error?: string; parsedObject?: unknown } {
+  let parsedObject;
+  let error = undefined;
+  try {
+    const parsedAny: unknown = JSON.parse(value);
+    if (Array.isArray(parsedAny)) {
+      error = "Request content must be an object, not an array";
+    } else if (parsedAny == null /* eslint-disable-line no-restricted-syntax */) {
+      error = "Request content must be an object, not null";
+    } else if (typeof parsedAny !== "object") {
+      error = `Request content must be an object, not ‘${typeof parsedAny}’`;
+    } else {
+      parsedObject = parsedAny;
+    }
+  } catch (e) {
+    error = value.length !== 0 ? e.message : "Enter valid request content as JSON";
+  }
+  return { error, parsedObject };
+}
+
+export function CallService({ context }: Props): JSX.Element {
+  // panel extensions must notify when they've completed rendering
+  // onRender will setRenderDone to a done callback which we can invoke after we've rendered
+  const [renderDone, setRenderDone] = useState<() => void>(() => () => {});
+  const [state, setState] = useState<State | undefined>();
+  const [config, setConfig] = useState<Config>(() => ({
+    ...defaultConfig,
+    ...(context.initialState as Partial<Config>),
+  }));
+  const { classes } = useStyles({ buttonColor: config.buttonColor });
+
+  useEffect(() => {
+    context.saveState(config);
+    context.setDefaultPanelTitle(
+      config.serviceName === "" ? undefined : `Call service ${config.serviceName}`,
+    );
+  }, [config, context]);
+
+  useEffect(() => {
+    context.onRender = (_renderState, done) => {
+      setRenderDone(() => done);
+    };
+
+    return () => {
+      context.onRender = undefined;
+    };
+  }, [context]);
+
+  const { error: requestParseError, parsedObject } = useMemo(
+    () => parseInput(config.requestPayload ?? ""),
+    [config.requestPayload],
+  );
+
+  const settingsActionHandler = useCallback(
+    (action: SettingsTreeAction) =>
+      setConfig((prevConfig) => settingsActionReducer(prevConfig, action)),
+    [setConfig],
+  );
+
+  const settingsTree = useSettingsTree(config);
+  useEffect(() => {
+    context.updatePanelSettingsEditor({
+      actionHandler: settingsActionHandler,
+      nodes: settingsTree,
+    });
+  }, [context, settingsActionHandler, settingsTree]);
+
+  const statusMessage = useMemo(() => {
+    if (context.callService == undefined) {
+      return "Connect to a data source that supports calling services";
+    }
+    if (!config.serviceName) {
+      return "Configure a service in the panel settings";
+    }
+    return undefined;
+  }, [context, config.serviceName]);
+
+  const canCallService = Boolean(
+    context.callService != undefined &&
+      config.requestPayload &&
+      config.serviceName &&
+      parsedObject != undefined &&
+      state?.status !== "requesting",
+  );
+
+  const callServiceClicked = useCallback(async () => {
+    if (!context.callService) {
+      setState({ status: "error", value: "The data source does not allow calling services" });
+      return;
+    }
+
+    try {
+      setState({ status: "requesting", value: `Calling ${config.serviceName}...` });
+      const response = await context.callService(
+        config.serviceName!,
+        JSON.parse(config.requestPayload!),
+      );
+      setState({ status: "success", value: JSON.stringify(response, undefined, 2) ?? "" });
+    } catch (err) {
+      setState({ status: "error", value: (err as Error).message });
+      log.error(err);
+    }
+  }, [context, config.serviceName, config.requestPayload]);
+
+  // Indicate render is complete - the effect runs after the dom is updated
+  useEffect(() => {
+    renderDone();
+  }, [renderDone]);
+
+  return (
+    <Stack fullHeight>
+      <Stack
+        flex="auto"
+        gap={1}
+        padding={1.5}
+        position="relative"
+        direction={config.layout === "horizontal" ? "row" : "column"}
+      >
+        <Stack flexGrow="1">
+          <Typography variant="caption" noWrap>
+            Request
+          </Typography>
+          <TextField
+            variant="outlined"
+            className={classes.textarea}
+            multiline
+            size="small"
+            placeholder="Enter service request as JSON"
+            value={config.requestPayload}
+            onChange={(event) => setConfig({ ...config, requestPayload: event.target.value })}
+            error={requestParseError != undefined}
+          />
+          {requestParseError && (
+            <Typography variant="caption" noWrap color={requestParseError ? "error" : undefined}>
+              {requestParseError}
+            </Typography>
+          )}
+        </Stack>
+        <Stack flexGrow="1">
+          <Typography variant="caption" noWrap>
+            Response
+          </Typography>
+          <TextField
+            variant="outlined"
+            className={classes.textarea}
+            multiline
+            size="small"
+            placeholder="Response"
+            value={state?.value}
+            error={state?.status === "error"}
+          />
+        </Stack>
+      </Stack>
+      <Stack
+        direction="column-reverse"
+        justifyContent="center"
+        alignItems="center"
+        overflow="hidden"
+        paddingBottom={1.5}
+      >
+        <Tooltip title={config.buttonTooltip}>
+          <span>
+            <Button
+              className={classes.button}
+              variant="contained"
+              disabled={!canCallService}
+              onClick={callServiceClicked}
+            >
+              {config.buttonText}
+            </Button>
+          </span>
+        </Tooltip>
+        {statusMessage && (
+          <Typography variant="caption" noWrap>
+            {statusMessage}
+          </Typography>
+        )}
+      </Stack>
+    </Stack>
+  );
+}
