+++ packages/studio-base/src/panels/CallService/settings.ts
@@ -3,18 +3,15 @@
 // file, You can obtain one at http://mozilla.org/MPL/2.0/
 
 import { produce } from "immer";
-import { isEqual, set } from "lodash";
-import { useCallback, useEffect, useMemo } from "react";
+import { set } from "lodash";
+import { useMemo } from "react";
 
-import { Immutable, SettingsTreeAction, SettingsTreeNodes } from "@foxglove/studio";
-import { usePanelSettingsTreeUpdate } from "@foxglove/studio-base/providers/PanelStateContextProvider";
-import { RosDatatypes } from "@foxglove/studio-base/types/RosDatatypes";
-import { SaveConfig } from "@foxglove/studio-base/types/panels";
-import buildSampleMessage from "@foxglove/studio-base/util/buildSampleMessage";
+import { useShallowMemo } from "@foxglove/hooks";
+import { SettingsTreeAction, SettingsTreeNodes } from "@foxglove/studio";
 
-import { CallServiceConfig } from "./types";
+import { Config } from "./types";
 
-export const defaultConfig: CallServiceConfig = {
+export const defaultConfig: Config = {
   requestPayload: "{}",
   layout: "vertical",
   buttonText: "Call service",
@@ -27,97 +24,47 @@ function serviceError(serviceName?: string) {
   return undefined;
 }
 
-const buildSettingsTree = (
-  config: CallServiceConfig,
-  schemaNames: string[],
-): SettingsTreeNodes => ({
-  general: {
-    fields: {
-      serviceName: {
-        label: "Service name",
-        input: "string",
-        error: serviceError(config.serviceName),
-        value: config.serviceName ?? "",
+export function settingsActionReducer(prevConfig: Config, action: SettingsTreeAction): Config {
+  return produce(prevConfig, (draft) => {
+    if (action.action === "update") {
+      const { path, value } = action.payload;
+      set(draft, path.slice(1), value);
+    }
+  });
+}
+
+export function useSettingsTree(config: Config): SettingsTreeNodes {
+  const settings = useMemo(
+    (): SettingsTreeNodes => ({
+      general: {
+        fields: {
+          serviceName: {
+            label: "Service name",
+            input: "string",
+            error: serviceError(config.serviceName),
+            value: config.serviceName ?? "",
+          },
+          layout: {
+            label: "Layout",
+            input: "toggle",
+            options: [
+              { label: "Vertical", value: "vertical" },
+              { label: "Horizontal", value: "horizontal" },
+            ],
+            value: config.layout ?? defaultConfig.layout,
+          },
+        },
       },
-      datatype: {
-        label: "Request schema",
-        input: "autocomplete",
-        help: "Optional request schema, used to pre-populate the request",
-        items: schemaNames,
-        value: config.requestSchemaName ?? "",
+      button: {
+        label: "Button",
+        fields: {
+          buttonText: { label: "Title", input: "string", value: config.buttonText },
+          buttonTooltip: { label: "Tooltip", input: "string", value: config.buttonTooltip },
+          buttonColor: { label: "Color", input: "rgb", value: config.buttonColor },
+        },
       },
-      layout: {
-        label: "Layout",
-        input: "toggle",
-        options: [
-          { label: "Vertical", value: "vertical" },
-          { label: "Horizontal", value: "horizontal" },
-        ],
-        value: config.layout ?? defaultConfig.layout,
-      },
-    },
-  },
-  button: {
-    label: "Button",
-    fields: {
-      buttonText: { label: "Title", input: "string", value: config.buttonText },
-      buttonTooltip: { label: "Tooltip", input: "string", value: config.buttonTooltip },
-      buttonColor: { label: "Color", input: "rgb", value: config.buttonColor },
-    },
-  },
-});
-
-const getSampleMessage = (
-  datatypes: Immutable<RosDatatypes>,
-  datatype?: string,
-): string | undefined => {
-  if (datatype == undefined) {
-    return undefined;
-  }
-  const sampleMessage = buildSampleMessage(datatypes, datatype);
-  return sampleMessage != undefined ? JSON.stringify(sampleMessage, undefined, 2) : "{}";
-};
-
-export function useCallServicePanelSettings(
-  config: CallServiceConfig,
-  saveConfig: SaveConfig<CallServiceConfig>,
-  datatypes: Immutable<RosDatatypes>,
-): void {
-  const updatePanelSettingsTree = usePanelSettingsTreeUpdate();
-  const schemaNames = useMemo(() => Array.from(datatypes.keys()).sort(), [datatypes]);
-
-  const actionHandler = useCallback(
-    (action: SettingsTreeAction) => {
-      if (action.action !== "update") {
-        return;
-      }
-      const { path, value, input } = action.payload;
-
-      saveConfig(
-        produce<CallServiceConfig>((draft) => {
-          if (input === "autocomplete") {
-            if (isEqual(path, ["general", "datatype"])) {
-              const sampleMessage = getSampleMessage(datatypes, value);
-
-              draft.requestSchemaName = value;
-
-              if (sampleMessage) {
-                draft.requestPayload = sampleMessage;
-              }
-            }
-          } else {
-            set(draft, path.slice(1), value);
-          }
-        }),
-      );
-    },
-    [datatypes, saveConfig],
+    }),
+    [config],
   );
-
-  useEffect(() => {
-    updatePanelSettingsTree({
-      actionHandler,
-      nodes: buildSettingsTree(config, schemaNames),
-    });
-  }, [actionHandler, config, schemaNames, updatePanelSettingsTree]);
+  return useShallowMemo(settings);
 }
