+++ packages/studio-base/src/panels/CallService/types.ts
@@ -2,9 +2,8 @@
 // License, v2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/
 
-export type CallServiceConfig = {
+export type Config = {
   serviceName?: string;
-  requestSchemaName?: string;
   requestPayload?: string;
   layout?: "vertical" | "horizontal";
   buttonText?: string;
