+++ packages/studio-base/src/panels/Publish/buildSampleMessage.ts
